Parse Header into kmb
  - Refactor kmb creation
  - Parse all header in header tag
  - Allow main to print table
Next:
  - segfault in draw function. draw by aread housl be draw by row,
    column
